chore (refs T22614): move MultistepNav to client/js.
